2023 day 6 answer two no pruning

+++ 2023/aoc2023.day06/domain/RaceState.cs
@@ -1,6 +1,6 @@
 ﻿namespace aoc2023.day06.domain
 {
-    internal record RaceState(int TimeButtonHeld, int DistanceTravelled, int TimeRemaining, int Speed)
+    internal record RaceState(long TimeButtonHeld, long DistanceTravelled, long TimeRemaining, long Speed)
     {
     }
 }
